+++ src/FormElement/ItemSetSelect.php
@@ -3,11 +3,16 @@
 namespace Search\FormElement;
 
 use Laminas\View\Renderer\PhpRenderer;
+use Omeka\Api\Manager as ApiManager;
 use Search\Api\Representation\SearchPageRepresentation;
+use Search\Feature\SummarizeQueryInterface;
 use Search\Query;
 
-class ItemSetSelect implements SearchFormElementInterface
+class ItemSetSelect implements SearchFormElementInterface, SummarizeQueryInterface
 {
+    protected $api;
+    protected $translator;
+
     public function getLabel(): string
     {
         return 'Item sets'; // @translate
@@ -43,5 +48,51 @@ class ItemSetSelect implements SearchFormElementInterface
         if (!empty($data['item_set_id'])) {
             $query->addFacetFilter($formElementData['field_name'], array_filter($data['item_set_id']));
         }
+    }
+
+    public function setApiManager(ApiManager $api): void
+    {
+        $this->api = $api;
+    }
+
+    public function getApiManager(): ApiManager
+    {
+        return $this->api;
+    }
+
+    public function setTranslator($translator): void
+    {
+        $this->translator = $translator;
+    }
+
+    public function getTranslator()
+    {
+        return $this->translator;
+    }
+
+    public function summarizeQuery($data, $page): array
+    {
+        $summary = [];
+        $apiManager = $this->getApiManager();
+        $translator = $this->getTranslator();
+
+        if (!empty($data['item_set_id'])) {
+            $titles = [];
+            foreach ($data['item_set_id'] as $id) {
+                try {
+                    $itemSet = $apiManager->read('item_sets', $id)->getContent();
+                    $titles[] = $itemSet->title();
+                } catch (\Omeka\Api\Exception\NotFoundException $e) {
+                }
+            }
+
+            if (!empty($titles)) {
+                $summary[] = [
+                    'name' => $translator->translate('Item sets'),
+                    'value' => implode(', ', $titles),
+                ];
+            }
+        }
+        return $summary;
     }
 }